requirements.txt
python==3.8.13
pytorch==1.12.0
paddlepaddle-gpu>=2.2.2
paddlespeech==0.2.0
torchaudio==0.12.0
pyyaml==6.0
pysoundfile==0.10.3
yaml==0.2.5
tqdm==4.64.0
nltk==3.7
numpy==1.21.5
librosa==0.9.1
visualdl
munch
pydub